push

+++ src/textrenderer/loadglyphs.cc
@@ -52,7 +52,7 @@ void TextRenderer::load_glyphs(string const &font, size_t size)
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x   
+                static_cast<GLuint>(face->glyph->advance.x)
             }
         });
     }
